refactor: 유효성 검사 관련 함수 리팩토링 (상세 내용은 하단에 기재)
- Car 객체에는 길이 제한 없이 이름을 부여할 수 있도록 변경
- 이름 길이 유효성 검사 기능은 RacingGame 객체에서 수행하도록 변경
- validate 함수 선언형으로 리팩토링
- Validator.js: Domain 디렉토리로 이동
