+++ docs/REQUIREMENTS.md
@@ -5,18 +5,16 @@
 ## 기능 요구 사항
 
 ##### 자동차
-
-- [x] 자동차에 5자 이하의 이름을 부여할 수 있다.
+- [x] 자동차에 이름을 부여할 수 있다.
 - [x] 자동차는 기존 위치에서 정해진 값만큼 전진할 수 있다.
 
 ##### 경주 게임
-
+- [x] 게임에 참여하는 자동차의 이름은 5자 이하여야 한다.
 - [x] 자동차 경주는 5회로 고정하여 진행한다.
 - [x] 0에서 9 사이에서 무작위로 얻은 값이 4 이상일 때 자동차를 전진시킬 수 있다.
 - [x] 자동차 경주 게임을 완료한 후 우승자를 선별한다. (우승자는 한 명 이상일 수 있다.)
 
 ##### 입출력
-
 - [x] 자동차는 쉼표를 기준으로 구분하여 입력받는다.
 - [x] 전진하는 자동차를 출력할 때 자동차 이름을 같이 출력한다.
 - [x] 우승자는 쉼표를 기준으로 구분하여 출력된다.
